+++ app/assets/javascript/controllers/index.js
