+++ app/assets/javascript/routes/App.js
@@ -1,0 +1,19 @@
+import React from "react";
+import { Link, Routes, Route } from "react-router-dom";
+import Greeting from "./greeting";
+
+
+function App() {
+  return (
+    <div>
+        <nav>
+            <Link to="/"></Link>
+        </nav>
+        <Routes>
+            <Route path="/" element={<Greeting />} />
+        </Routes>
+
+    </div>
+  );
+}
+export default App;
