+++ app/controllers/greetings_controller.rb
@@ -1,0 +1,4 @@
+class GreetingsController < ApplicationController
+    def index
+    end
+end
