--- app/javascript/application.js
@@ -1,3 +1,0 @@
-// Entry point for the build script in your package.json
-import "@hotwired/turbo-rails";
-import "./controllers";
