--- app/javascript/components/greeting.js
@@ -1,8 +1,0 @@
-import React from "react";
-
-const Greeting = () => (
-    <div>
-        <h1>Hello!</h1>
-    </div>
-);
-export default Greeting;
