--- app/javascript/packs/Index.js
@@ -1,18 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { BrowserRouter as Router } from "react-router-dom";
-import App from "./routes/App";
-import store from "../Redux/ConfigureStore";
-import { Provider } from "react-redux";
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <Router>
-      <Provider store={store}>
-        <App />
-      </Provider>
-    </Router>
-  </React.StrictMode>
-);
-
