+++ app/views/greetings/index.html.erb
@@ -1,1 +1,1 @@
-<div id="root">
+<div id="root"></div>
