+++ config/routes.rb
@@ -4,8 +4,11 @@ Rails.application.routes.draw do
       resources :greetings, only: [:index]
     end
   end
+
+  root to: 'greetings#index'
+
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
-  # Defines the root path route ("/")
+  # Defines the root path route ("/")z
   # root "articles#index"
 end
