+++ package.json
@@ -1,6 +1,15 @@
 {
   "name": "app",
   "private": "true",
+  "scripts": {
+    "lint": "eslint --ext .js app/javascript",
+    "lint:fix": "eslint --fix --ext .js app/javascript",
+    "test:coverage": "jest --coverage",
+    "test": "jest --coverage --watch",
+    "start": "bin/dev",
+    "build": "webpack --config webpack.config.js",
+    "watch": "webpack --config webpack.config.js --watch --mode development"
+  },
   "dependencies": {
     "@babel/core": "^7.19.0",
     "@babel/preset-env": "^7.19.0",
@@ -14,6 +23,8 @@
     "react-router-dom": "^6.3.0",
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.4.1"
+    "redux-thunk": "^2.4.1",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.9.2"
   }
 }
