+++ webpack.config.js
@@ -5,7 +5,7 @@ module.exports = {
   mode: "production",
   devtool: "source-map",
   entry: {
-    application: "./app/javascript/application.js"
+    application: "./app/assets/javascript/index.js"
   },
   module: {
     rules: [
